fix: validate float inputs, guard integer overflow, and harden contrib integrations
Guard NaN/Infinity/negative floats before casting to i64 in delay_seconds,
expires, retry_backoff, and retry_delays. Add progress 0-100 range check.
Fix Django admin page param parsing, singleton race condition, and broad
exception catch. Refresh stale job status in FastAPI after aresult()
timeout. Use per-signature timeouts in canvas group/chord. Fix Redis
dequeue_score i32 overflow. Log cascade_cancel failures in DLQ. Validate
worker payload tuple length. Narrow version import exception.

+++ crates/taskito-core/src/storage/redis_backend/jobs.rs
@@ -9,7 +9,7 @@ use crate::storage::QueueStats;
 /// Compute dequeue score: higher priority → lower score → dequeued first.
 /// Within same priority, earlier scheduled_at wins.
 fn dequeue_score(priority: i32, scheduled_at: i64) -> f64 {
-    (1000 - priority) as f64 * 10_000_000_000_000.0 + scheduled_at as f64
+    (1000i64 - priority as i64) as f64 * 10_000_000_000_000.0 + scheduled_at as f64
 }
 
 impl RedisStorage {
